@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 /**
  * The {@code AccountEndpoint} interface defines the exposed restful services in the account bounded context.
